Refactor select callback into manager
