+++ src/hooks/index.ts
@@ -3,7 +3,6 @@ export * from './useInputSizer'
 export * from './useCaptureFocus'
 export * from './useListBox'
 export * from './useManagerTwo'
-export * from './useOnSelect'
 export * from './useOption'
 export * from './usePublicAPI'
 export * from './useRoot'
